requirements.txt
django~=4.1
djangorestframework~=3.14.0
certifi==2022.12.7
gunicorn==20.1.0
asgiref
pytz
sqlparse
pillow